+++ packages/elements/src/components/ui/badge/Badge.tsx
@@ -1,35 +1,22 @@
-import { Box, SmallText } from "@stenajs-webui/core";
+import { SmallText } from "@stenajs-webui/core";
 import * as React from "react";
 import styles from "./Badge.module.css";
 
+export type BadgeType = "notification" | "warning" | "error";
+
 export interface BadgeProps {
-  color?: string;
-  textColor?: string;
   label?: string | number;
-  size?: string | number;
+  type?: BadgeType;
 }
 
 export const Badge: React.FC<BadgeProps> = React.memo(
-  ({
-    textColor = "var(--swui-badge-text-color)",
-    color = "var(--swui-badge-bg-color)",
-    size = "var(--swui-badge-size)",
-    label
-  }) => {
+  ({ label, type = "notification" }) => {
+    const className = styles.badge + " " + styles[type];
+
     return (
-      <Box
-        className={styles.badge}
-        overflow={"hidden"}
-        borderRadius={"50%"}
-        borderWidth={"0px"}
-        background={color}
-        width={size}
-        height={size}
-        alignItems={"center"}
-        justifyContent={"center"}
-      >
-        <SmallText color={textColor}>{label}</SmallText>
-      </Box>
+      <div className={className}>
+        <SmallText color={"var(--swui-badge-text-color)"}>{label}</SmallText>
+      </div>
     );
   }
 );