+++ packages/elements/src/components/ui/chip/Chip.stories.tsx
@@ -1,16 +1,37 @@
-import { Chip } from "@stenajs-webui/elements";
+import { Column, LargeText, Row, Space } from "@stenajs-webui/core";
+import { Chip, ChipVariant } from "@stenajs-webui/elements";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
 storiesOf("elements/Chip", module).add("standard", () => (
-  <div className={"margin-wrapper"}>
-    <Chip label={"Default"} />
-    <Chip label={"With onClick"} onClick={() => alert("hello world")} />
-    <Chip label={"With onRemove"} onRemove={() => alert("remove")} />
-    <Chip
-      label={"With both"}
-      onClick={() => alert("hello world")}
-      onRemove={() => alert("remove")}
-    />
-  </div>
+  <Column>
+    {(["primary", "secondary"] as Array<ChipVariant>).map(variant => (
+      <>
+        <LargeText>{variant}</LargeText>
+        <Space />
+        <Row>
+          <div className={"indent-items"}>
+            <Chip variant={variant} label={"Default"} />
+            <Chip
+              variant={variant}
+              label={"With onClick"}
+              onClick={() => alert("hello world")}
+            />
+            <Chip
+              variant={variant}
+              label={"With onClickRemove"}
+              onClickRemove={() => alert("remove")}
+            />
+            <Chip
+              variant={variant}
+              label={"With both"}
+              onClick={() => alert("hello world")}
+              onClickRemove={() => alert("remove")}
+            />
+          </div>
+        </Row>
+        <Space num={3} />
+      </>
+    ))}
+  </Column>
 ));