+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Next release
 
-* Coming soon...
+### Breaking changes
+
+* Badge props changed. `color` and `textColor` have been removed, and `type` has been added.
 
 ## 0.0.51
 
@@ -12,11 +14,11 @@
 
 ## 0.0.50
 
-* Fix typescript errors in calendar package.
+* Fix Typescript errors in calendar package.
 
 ## 0.0.49
 
-* Modals can now be defined as draggable!
+* Modals are now draggable. Off by default, enable with prop `draggable`.
 
 ### Fixes