+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Next release
 
+### PSA
+
+* We are moving from React context based theme to CSS variables.
+* We are also starting to implement components using HTML elements and CSS modules, instead of `core` components.
+
 ### Breaking changes
 
 * Badge props changed. `color` and `textColor` have been removed, and `type` has been added.
+* Badge now uses CSS vars.
 
 ## 0.0.51